In `bp_pre_schema_upgrade()`, only drop indexes if the table exists.
See [9716]. See #6346.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@9717 cdf35c40-ae34-48e0-9cc9-0c9da1808c22

+++ src/bp-core/bp-core-update.php
@@ -286,7 +286,9 @@ function bp_pre_schema_upgrade() {
 
 		foreach ( $tables as $table_name => $indexes ) {
 			foreach ( $indexes as $index ) {
-				$wpdb->query( "ALTER TABLE $table_name DROP INDEX $index" );
+				if ( $wpdb->query( $wpdb->prepare( "SHOW TABLES LIKE %s", $wpdb->esc_like( $table_name ) ) ) ) {
+					$wpdb->query( "ALTER TABLE $table_name DROP INDEX $index" );
+				}
 			}
 		}
 	}
